[BUGFIX] [Hacker Rank]: Project Euler #1: Multiples of 3 and 5. Types changed to reach problem limits. Solved ✓

+++ src/lib/exercises/include/exercises/hackerrank/projecteuler/euler001.hpp
@@ -2,6 +2,6 @@
 
 namespace hackerrank::projecteuler {
 
-int euler001(int a, int b, int n);
+unsigned long euler001(int a, int b, int n);
 
 }
